apply visitor
