+++ eng/packages/http-client-csharp/generator/Azure.Generator/src/AzureClientPlugin.cs
@@ -52,13 +52,10 @@ public class AzureClientPlugin : ClientModelPlugin
         AddMetadataReference(MetadataReference.CreateFromFile(typeof(Response).Assembly.Location));
         var sharedSourceDirectory = Path.Combine(Path.GetDirectoryName(typeof(AzureClientPlugin).Assembly.Location)!, "Shared", "Core");
         AddSharedSourceDirectory(sharedSourceDirectory);
+        AddVisitor(new NamespaceVisitor());
         if (IsAzureArm.Value)
         {
-            AddVisitor(new AzureArmVisitor());
-        }
-        else
-        {
-            AddVisitor(new AzureVisitor());
+            AddVisitor(new RestClientVisitor());
         }
     }
