+++ eng/packages/http-client-csharp/generator/Azure.Generator/src/RestClientVisitor.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using Microsoft.TypeSpec.Generator.ClientModel;
 using Microsoft.TypeSpec.Generator.ClientModel.Providers;
 using Microsoft.TypeSpec.Generator.Primitives;
 using Microsoft.TypeSpec.Generator.Providers;
@@ -9,7 +10,7 @@ using System.IO;
 namespace Azure.Generator
 {
     // only apply for MPG
-    internal class AzureArmVisitor : AzureVisitor
+    internal class RestClientVisitor : ScmLibraryVisitor
     {
         /// <inheritdoc/>
         protected override TypeProvider? Visit(TypeProvider type)
